+++ internal/mdathome/structs.go
@@ -25,6 +25,7 @@ type ClientSettings struct {
 	OverrideUpstream        string `json:"override_upstream"`
 	RejectInvalidTokens     bool   `json:"reject_invalid_tokens"`
 	VerifyImageIntegrity    bool   `json:"verify_image_integrity"`
+	LowMemoryMode           bool   `json:"low_memory_mode"`
 
 	LogLevel              string `json:"log_level"`
 	MaxLogSizeInMebibytes int    `json:"max_log_size_in_mebibytes"`