+++ internal/mdathome/constants.go
@@ -8,7 +8,7 @@ package mdathome
 var ClientVersion string
 
 // ClientSpecification is the integer version of the official specification this client is supposed to work against
-const ClientSpecification int = 20
+const ClientSpecification int = 23
 
 // Backend settings (Swap to use testnet)
 const apiBackend string = "https://api.mangadex.network" // "https://mangadex-test.net"