+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - [2021-01-21] Adjusted logging of diskcache by [@lflare].
+- [2021-01-23] Made server read/write timeouts more aggresive with 30s and 1m respectively by [@lflare].
 
 ### Deprecated