+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [2021-01-21] Adjusted logging of diskcache by [@lflare].
 - [2021-01-23] Made server read/write timeouts more aggresive with 30s and 1m respectively by [@lflare].
 - [2021-01-23] Properly pre-processed IP address to only log IP addresses without ports by [@lflare].
+- [2021-01-24] Reworked for auto-downloading of MaxMind databases for geolocation by [@lflare].
 
 ### Deprecated