+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - [2021-01-28] Reworked diskcache to stream files more efficiently by [@lflare].
+- [2021-02-02] Added `disable_tokens` handling from backend by [@lflare].
+- [2021-02-02] Updated to client specification 23 by [@lflare].
 
 ### Deprecated